feat(i18n): refactor name and move .properties and interface file to studio.client.application
