+++ src/gwt/src/org/rstudio/studio/client/application/StudioClientApplicationConstants_en.properties
